requirements.txt
black
flake8
hurry.filesize==0.9
pytest
spikes